chore(core/rust): move micropython out to separate crate
