+++ core/embed/micropython/inc
@@ -1,0 +1,1 @@
+../../../../vendor/micropython
