+++ core/embed/micropython/micropython.h
@@ -1,0 +1,8 @@
+#include "py/gc.h"
+#include "py/mphal.h"
+#include "py/nlr.h"
+#include "py/obj.h"
+#include "py/objstr.h"
+#include "py/runtime.h"
+
+#include "../../../upymod/trezorobj.h"
