+++ core/embed/micropython/mpconfigport.h
@@ -1,0 +1,1 @@
+../../../projects/unix/mpconfigport.h
