+++ core/embed/micropython/mphalport.h
@@ -1,0 +1,1 @@
+../../../projects/firmware/mphalport.h
