+++ core/embed/micropython/rustfmt.toml
@@ -1,0 +1,3 @@
+wrap_comments = true
+group_imports = "StdExternalCrate"
+imports_granularity = "Module"
