+++ core/embed/micropython/src/iter.rs
