+++ core/embed/micropython/src/lib.rs
@@ -19,8 +19,5 @@ pub mod simple_type;
 pub mod typ;
 pub mod util;
 
-#[cfg(feature = "debug")]
-pub mod logging;
-
 #[cfg(test)]
 pub mod testutil;
