+++ core/embed/micropython/src/list.rs
