+++ core/embed/micropython/src/macros.rs
