+++ core/embed/micropython/src/qstr.rs
