+++ core/embed/micropython/src/runtime.rs
