+++ core/embed/rust/Cargo.toml
@@ -14,7 +14,7 @@ layout_bolt = []
 layout_caesar = []
 layout_delizia = []
 layout_eckhart = []
-micropython = []
+micropython = ["dep:micropython"]
 protobuf = ["micropython"]
 ui = []
 dma2d = []
@@ -122,6 +122,7 @@ spin = { version = "0.9.8", features = ["rwlock"], default-features = false }
 trezor-tjpgdec = { version = "0.1.0", path = "../../../rust/trezor-tjpgdec" }
 ufmt = "0.2.0"
 zeroize = { version = "1.8.2", default-features = false, optional = true }
+micropython = { version = "0.1.0", path = "../micropython", optional = true }
 
 # Runtime dependencies
